Move methods to Seats class, fix docs, remove empty class.
